requirements.txt
portkey-ai>=1.0.0
python-dotenv>=0.19.0
rich>=12.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0